requirements.txt
PyQt6>=6.6.0
requests>=2.31.0
keyboard>=0.13.5
pygame>=2.5.0
# Stretch goal — OCR item scanner (install separately if desired)
mss>=9.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
